+++ src/Shapes/Obj.hpp
@@ -19,6 +19,6 @@ namespace RayTracer {
 
             ~Obj() = default;
 
-            bool hit(const Math::Ray3D &ray, float tmin, float tmax, hits &hit) const override;
+            bool hit(const Math::Ray3D &ray, double tmin, double tmax, hits &hit) const override;
     };
 }
