+++ src/Shapes/Prism.cpp
@@ -11,7 +11,7 @@
 RayTracer::Prism::Prism() {}
 
 
-bool RayTracer::Prism::hit(const Math::Ray3D &ray, float tmin, float tmax, hits &hit) const
+bool RayTracer::Prism::hit(const Math::Ray3D &ray, double tmin, double tmax, hits &hit) const
 {
     (void)ray;
     (void)tmin;
